improve search by english titles
getHeader() removes auxiliary words like "to", "and", "may", ... with removeAuxiliary(). This is desirable if a translation is created to be searched with, but not if the text is already English, as API and search with the exact stackoverflow title will give better results. Therefore, I replaced the getHeader() call when the text is already English with textContent().
getHeader() has a feature that if passed as an array, the parameter is not understood as a selector, but as text. I implemented this feature also for textContent() to keep the compatibility, because byHeader() is called multiple times with the use of this feature.
All in all, this improves the performance with English titles noticeably.